Refactor to move `make_search_context_tool` from `tools.py` to `langchain_tools.py` and update imports across executors.
