+++ databao/agent/duckdb/react_tools.py
@@ -12,7 +12,7 @@ from pydantic import BaseModel
 
 from databao.agent.core import Domain
 from databao.agent.duckdb.schema_inspection import inspect_duckdb_schema, summarize_duckdb_schema
-from databao.agent.executors.tools import make_search_context_tool
+from databao.agent.executors.langchain_tools import make_search_context_tool
 
 _LOGGER = logging.getLogger(__name__)
