+++ databao/agent/executors/dbt/graph.py
@@ -30,7 +30,7 @@ from databao.agent.executors.dbt.dbt_runner import (
 from databao.agent.executors.dbt.query_runner import QueryRunnerFactory
 from databao.agent.executors.llm import chat, model_bind_tools
 from databao.agent.executors.query_expansion import QueryExpansionConfig
-from databao.agent.executors.tools import make_search_context_tool
+from databao.agent.executors.langchain_tools import make_search_context_tool
 
 
 @dataclass(frozen=True)
